chore(mocks/geolocationServices.mock.js): Function for geolocation services mocked
