+++ packages/api-gateway/__tests__/geolocation.test.js
@@ -21,7 +21,7 @@ describe("geolocation/address endpoint", () => {
         }),
       })
     );
-  });
+  }, 10000);
 
   test("Given wrong or null latitude and longitude when a user select a mark in a map, then return status code 400 ", async () => {
     const address = await api.inject({
@@ -30,5 +30,5 @@ describe("geolocation/address endpoint", () => {
     });
 
     expect(address.statusCode).toEqual(400);
-  });
+  }, 10000);
 });
